rustc_thread_pool

Module join

Functions

join

Takes two closures and potentially runs them in parallel. It returns a pair of the results from those closures.

join_context

Identical to join, except that the closures have a parameter that provides context for the way the closure has been called, especially indicating whether they’re executing on a different thread than where join_context was called. This will occur if the second job is stolen by a different thread, or if join_context was called from outside the thread pool to begin with.